Fixed issue where time would not progress properly in animations when the game framerate was significantly high (greater than several hundred FPS).
Changed the method signatures of Game::getAbsoluteTime, Game::getGameTime, Game::update and Game::render to use floating point instead of integers in order to compute fractions of milliseconds of elapsed time (to support the above fix).

+++ gameplay/src/AnimationController.cpp
@@ -93,7 +93,7 @@ void AnimationController::unschedule(AnimationClip* clip)
         _state = IDLE;
 }
 
-void AnimationController::update(long elapsedTime)
+void AnimationController::update(float elapsedTime)
 {
     if (_state != RUNNING)
         return;
